Use shared logic to setup tests
